updated tests in new folder an broke out into seperate files, renamed user route files, alphabatized variables, added post delete put for users, removed Number, added merchant put with test
