+++ server/api/brandRoutes.js
@@ -1,5 +1,5 @@
-const router = require('express').Router()
 const {Brand, Product} = require('../db/models')
+const router = require('express').Router()
 module.exports = router
 
 router.get('/', async (req, res, next) => {
@@ -12,7 +12,7 @@ router.get('/', async (req, res, next) => {
 })
 
 router.get('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
+  const id = req.params.id
   try {
     const brand = await Brand.findByPk(id, {
       include: [
@@ -37,7 +37,7 @@ router.post('/', async (req, res, next) => {
 })
 
 router.put('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
+  const id = req.params.id
   try {
     const brand = await Brand.findByPk(id)
     brand.update(req.body)
@@ -48,7 +48,7 @@ router.put('/:id', async (req, res, next) => {
 })
 
 router.delete('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
+  const id = req.params.id
   try {
     const brand = await Brand.findByPk(id)
     brand.destroy()
