+++ server/api/reviewRoutes.js
@@ -21,7 +21,7 @@ router.post('/', async (req, res, next) => {
 })
 
 router.put('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
+  const id = req.params.id
   try {
     const review = await Review.findByPk(id)
     review.update(req.body)
@@ -32,7 +32,7 @@ router.put('/:id', async (req, res, next) => {
 })
 
 router.delete('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
+  const id = req.params.id
   try {
     const review = await Review.findByPk(id)
     review.destroy()
