--- server/api/users.js
@@ -1,31 +1,0 @@
-const router = require('express').Router()
-const {User, Review} = require('../db/models')
-module.exports = router
-
-router.get('/', async (req, res, next) => {
-  try {
-    const users = await User.findAll({
-      attributes: ['id', 'email']
-    })
-    res.json(users)
-  } catch (err) {
-    next(err)
-  }
-})
-
-router.get('/:id', async (req, res, next) => {
-  const id = Number(req.params.id)
-  try {
-    const users = await User.findByPk(id, {
-      include: [
-        {
-          model: Review
-        }
-      ],
-      attributes: ['id', 'email']
-    })
-    res.json(users)
-  } catch (err) {
-    next(err)
-  }
-})
