+++ server/db/models/index.js
@@ -1,9 +1,9 @@
-const User = require('./user')
-const Product = require('./product')
-const Review = require('./review')
-const Department = require('./department')
 const Brand = require('./brand')
 const {CreditCard, Merchant} = require('./creditCards')
+const Department = require('./department')
+const Product = require('./product')
+const Review = require('./review')
+const User = require('./user')
 
 CreditCard.belongsTo(User)
 CreditCard.belongsTo(Merchant)
